Extract metaball edge advanced top bar
